PCLOUDS-1602: SFM refactor - fixes
